The game now correctly reads the json file. The only thing left to do is to finish the game engine.
Also, I would like to implement more flavor text into the game. Specifically, the fight should play out differently depending on what items the player has. This is both for enhancing enjoyment, but also to let the player know what they were missing if they died.

Also, there should be a short delay between steps so that the player can read. Maybe even waiting for the player to press enter so that the text advances.